@@ -11,8 +11,8 @@ part of cool_ui;
 /// Widget Util.
 class _WidgetUtil {
   bool _hasMeasured = false;
-  double _width;
-  double _height;
+  double? _width;
+  double? _height;
 
   /// Widget rendering listener.
   /// Widget渲染监听.
@@ -22,8 +22,8 @@ class _WidgetUtil {
   void asyncPrepare(
       BuildContext context, bool isOnce, ValueChanged<Rect> onCallBack) {
     if (_hasMeasured) return;
-    WidgetsBinding.instance.addPostFrameCallback((Duration timeStamp) {
-      RenderBox box = context.findRenderObject();
+    WidgetsBinding.instance!.addPostFrameCallback((Duration timeStamp) {
+      RenderBox? box = context.findRenderObject() as RenderBox?;
       if (box != null && box.semanticBounds != null) {
         if (isOnce) _hasMeasured = true;
         double width = box.semanticBounds.width;
@@ -38,9 +38,9 @@ class _WidgetUtil {
   }
 
   /// Widget渲染监听.
-  void asyncPrepares(bool isOnce, ValueChanged<Rect> onCallBack) {
+  void asyncPrepares(bool isOnce, ValueChanged<Rect?>? onCallBack) {
     if (_hasMeasured) return;
-    WidgetsBinding.instance.addPostFrameCallback((Duration timeStamp) {
+    WidgetsBinding.instance!.addPostFrameCallback((Duration timeStamp) {
       if (isOnce) _hasMeasured = true;
       if (onCallBack != null) onCallBack(null);
     });
@@ -49,8 +49,8 @@ class _WidgetUtil {
   ///get Widget Bounds (width, height, left, top, right, bottom and so on).Widgets must be rendered completely.
   ///获取widget Rect
   static Rect getWidgetBounds(BuildContext context) {
-    RenderBox box = context.findRenderObject();
-    return (box != null && box.semanticBounds != null)
+    RenderBox? box = context.findRenderObject() as RenderBox?;
+    return (box != null)
         ? box.semanticBounds
         : Rect.zero;
   }
@@ -58,7 +58,7 @@ class _WidgetUtil {
   ///Get the coordinates of the widget on the screen.Widgets must be rendered completely.
   ///获取widget在屏幕上的坐标,widget必须渲染完成
   static Offset getWidgetLocalToGlobal(BuildContext context) {
-    RenderBox box = context.findRenderObject();
+    RenderBox? box = context.findRenderObject() as RenderBox?;
     return box == null ? Offset.zero : box.localToGlobal(Offset.zero);
   }
 }
