@@ -7,7 +7,7 @@ class NumberKeyboard extends StatelessWidget{
     return mediaQuery.size.width / 3 / 2 * 4;
   }
   final KeyboardController controller ;
-  const NumberKeyboard({this.controller});
+  const NumberKeyboard({required this.controller});
 
   static register(){
     CoolKeyboard.addKeyboard(NumberKeyboard.inputType,KeyboardConfig(builder: (context,controller, params){
@@ -68,17 +68,14 @@ class NumberKeyboard extends StatelessWidget{
     );
   }
 
-  Widget buildButton(String title,{String value}){
-    if(value == null){
-      value = title;
-    }
+  Widget buildButton(String title,{String? value}){
     return Container(
       color: Colors.white,
       child: GestureDetector(
         behavior: HitTestBehavior.translucent,
         child: Center(child: Text(title),),
         onTap: (){
-          controller.addText(value);
+          controller.addText(value ?? title);
         },
       ),
     );
